Add PiTFT buttons and backlight dimming
Added the beginning of the code that will be used for utilising the
PiTFT buttons on GPIO 17,22,23,27. For now button 4 (GPIO 17) manually
toggles the PTII relay.
Added code to dim the backlight of the PiTFT screen when the PTII relay
turns the lamp off at night to prevent it turning into a spotlight at
night.
Tidied up code to read a little better.